+++ src/Core/Language.Tests/Parser/SchemaParserTests.cs
@@ -1,75 +1,130 @@
-﻿using Xunit;
+﻿using Snapshooter.Xunit;
+using Xunit;
 
 namespace HotChocolate.Language
 {
     public class SchemaParserTests
     {
-        [Fact]
+         [Fact]
         public void ParserSimpleObjectType()
         {
             // arrange
-            string sourceText = "type a { b: String c: Int }";
+            string sourceText = "type a @foo(a: \"123\") " +
+                "{ b: String @foo(a: \"123\") " +
+                "c(d: F = ENUMVALUE @foo(a: \"123\")): Int }";
+            var parser = new Parser();
 
             // act
-            var parser = new Parser();
             DocumentNode document = parser.Parse(sourceText);
 
             // assert
-            Assert.Collection(document.Definitions,
-                t =>
-                {
-                    Assert.IsType<ObjectTypeDefinitionNode>(t);
-                    var objectTypeDefinition = (ObjectTypeDefinitionNode)t;
-                    Assert.Equal(NodeKind.ObjectTypeDefinition, objectTypeDefinition.Kind);
-                    Assert.Equal("a", objectTypeDefinition.Name.Value);
-                    Assert.Collection(objectTypeDefinition.Fields,
-                        f =>
-                        {
-                            Assert.Equal("b", f.Name.Value);
-                            Assert.IsType<NamedTypeNode>(f.Type);
-                            Assert.Equal("String", ((NamedTypeNode)f.Type).Name.Value);
-                        },
-                        f =>
-                        {
-                            Assert.Equal("c", f.Name.Value);
-                            Assert.IsType<NamedTypeNode>(f.Type);
-                            Assert.Equal("Int", ((NamedTypeNode)f.Type).Name.Value);
-                        });
-                });
+            SchemaSyntaxSerializer.Serialize(document).MatchSnapshot();
+        }
+
+        [Fact]
+        public void ParserInputObjectType()
+        {
+            // arrange
+            string sourceText = "input a @foo(a: \"123\") " +
+                "{ b: String @foo(a: \"123\") c: Int = 123 }";
+            var parser = new Parser();
+
+            // act
+            DocumentNode document = parser.Parse(sourceText);
+
+            // assert
+            SchemaSyntaxSerializer.Serialize(document).MatchSnapshot();
+        }
+
+        [Fact]
+        public void ParserScalarType()
+        {
+            // arrange
+            string sourceText = "scalar FOO @foo(a: \"123\")";
+            var parser = new Parser();
+
+            // act
+            DocumentNode document = parser.Parse(sourceText);
+
+            // assert
+            SchemaSyntaxSerializer.Serialize(document).MatchSnapshot();
         }
 
         [Fact]
         public void ParserSimpleInterfaceType()
         {
             // arrange
-            string sourceText = "interface a { b: String c: Int }";
+            string sourceText = "interface a @foo(a: \"123\") " +
+                "{ b: String @foo(a: \"123\") " +
+                "c(d: F = ENUMVALUE @foo(a: \"123\")): Int }";
+             var parser = new Parser();
 
             // act
-            var parser = new Parser();
             DocumentNode document = parser.Parse(sourceText);
 
             // assert
-            Assert.Collection(document.Definitions,
-                t =>
-                {
-                    Assert.IsType<InterfaceTypeDefinitionNode>(t);
-                    var objectTypeDefinition = (InterfaceTypeDefinitionNode)t;
-                    Assert.Equal(NodeKind.InterfaceTypeDefinition, objectTypeDefinition.Kind);
-                    Assert.Equal("a", objectTypeDefinition.Name.Value);
-                    Assert.Collection(objectTypeDefinition.Fields,
-                        f =>
-                        {
-                            Assert.Equal("b", f.Name.Value);
-                            Assert.IsType<NamedTypeNode>(f.Type);
-                            Assert.Equal("String", ((NamedTypeNode)f.Type).Name.Value);
-                        },
-                        f =>
-                        {
-                            Assert.Equal("c", f.Name.Value);
-                            Assert.IsType<NamedTypeNode>(f.Type);
-                            Assert.Equal("Int", ((NamedTypeNode)f.Type).Name.Value);
-                        });
-                });
+            SchemaSyntaxSerializer.Serialize(document).MatchSnapshot();
+        }
+
+        [Fact]
+        public void ParseEnum()
+        {
+            // arrange
+            string sourceText = "enum Foo @foo(a: \"123\") "
+                + "{ BAR @foo(a: 123) , BAZ }";
+            var parser = new Parser();
+
+            // act
+            DocumentNode document = parser.Parse(sourceText);
+
+            // assert
+            SchemaSyntaxSerializer.Serialize(document).MatchSnapshot();
+        }
+
+        [Fact]
+        public void ParseUnion()
+        {
+            // arrange
+            string sourceText = "union Foo @foo(a: \"123\") = "
+                + "BAR | BAZ ";
+            var parser = new Parser();
+
+            // act
+            DocumentNode document = parser.Parse(sourceText);
+
+            // assert
+            SchemaSyntaxSerializer.Serialize(document).MatchSnapshot();
+        }
+
+        [Fact]
+        public void ParseUnion_LeadingPipe()
+        {
+            // arrange
+            string sourceText = "union Foo @foo(a: \"123\") = "
+                + "| BAR | BAZ ";
+            var parser = new Parser();
+
+            // act
+            DocumentNode document = parser.Parse(sourceText);
+
+            // assert
+            SchemaSyntaxSerializer.Serialize(document).MatchSnapshot();
+        }
+
+        [Fact]
+        public void ParseSchemaDefinition()
+        {
+            // arrange
+            string sourceText = "\"\"\"\nDescription\n\"\"\"" +
+                "schema @foo(a: \"123\") " +
+                "{ query: Foo mutation: Bar subscription: Baz }";
+            var parser = new Parser();
+
+            // act
+            DocumentNode document = parser.Parse(sourceText);
+
+            // assert
+            SchemaSyntaxSerializer.Serialize(document).MatchSnapshot();
         }
     }
 }